min fix

+++ src/routes/Twitter.ts
@@ -263,7 +263,7 @@ export async function tipuser(twitter_user, action, id = '', amount, coin) {
                     if(balance > amount){
                         console.log('SENDING TO ADDRESS ' + pubAddr + ' ' + amount + ' ' + coin)
                         if(testmode === false){
-                            wallet.request('sendtoaddress',[pubAddr,parseInt(amount)]).then(function(txid){
+                            wallet.request('sendtoaddress',[pubAddr,parseFloat(amount)]).then(function(txid){
                                 message(
                                     twitter_user,
                                     "I've sent " + amount + " $" + coin + " to you! Check your TXID: " + txid['result'] + "!"
